requirements.txt
fastapi
hypercorn
google-cloud-bigquery